Change directory of admin API unit tests
